feat: add esm.mjs solution
